+++ config/routes.rb
@@ -1,2 +1,4 @@
 # Plugin's routes
 # See: http://guides.rubyonrails.org/routing.html
+
+get 'garfield/:name', to: 'garfield#show'