+++ test/functional/wiki_controller_test.rb
@@ -184,6 +184,17 @@ class WikiControllerTest < ActionController::TestCase
     assert_select 'script[src=?]', 'https://gist.github.com/plentz/6737338.js'
   end
 
+  def test_show_with_garfield_macro
+    @request.session[:user_id] = 1
+    @page.content.text = '{{garfield(2015,10,20)}}'
+    @page.content.save!
+    get :show, project_id: 1, id: @page_name
+    assert_response :success
+    assert_template 'show'
+    assert_select 'div.wiki img.garfield'
+    assert_select 'img[src=?]', '/garfield/ga151020/jpg'
+  end
+
   def test_show_with_weeknumber_macro
     @request.session[:user_id] = 1
     @page.content.text = '{{current_weeknumber}}'